improve applications management

+++ backend/src/server.py
@@ -9,7 +9,7 @@ import time
 
 from bottle import run, response, Bottle, error, static_file, json_dumps
 
-from utils.applications import get_applications
+from utils.applications import get_applications, APP_SCRIPT_STOP_SUFFIX, APP_SCRIPT_START_SUFFIX, APP_SCRIPT_EXT
 from utils.host import get_cube_ip
 
 verbose = False
@@ -70,76 +70,116 @@ def page_not_found(e):
 #     return json.dumps(files)
 
 
-@app.route('/api/start/<category>/<script>')
-def start_script(category, script):
+@app.route('/api/start/<category>/<application>')
+def start_application(category, application):
     """
     https://stackoverflow.com/questions/4789837/how-to-terminate-a-python-subprocess-launched-with-shell-true/4791612#4791612
 
+    For each application, two scripts must exist :
+
+    - <category>/<application>_start.sh
+    - <category>/<application>_stop.sh
+
     :param category:
-    :param script: the filename of the script, without the path
+    :param application: the name of the application
     :return:
     """
     global running_script
+
+    logging.debug(f'start_application({category}, {application})')
+
+    if category == 'special' and application == 'server_restart':
+        return server_restart()
+
     # If a script is already running, we stop it
     # TODO: add option to ignore request if a script is already running
     if running_script:
-        # If the requested script is already running, we do nothing
-        if running_script['category'] == category and running_script['script'] == script:
-            return {"message": f"{category}/{script} is already running"}
 
-        # response.status = 409   # conflict
-        # return {"error": f"A script is already running."}
+        logging.debug(f"{running_script['category']}/{running_script['application']} is already running")
+
+        # If the requested script is already running, we do nothing
+        # if running_script['category'] == category and running_script['application'] == application:
+        #     return {"message": f"{category}/{application} is already running"}
+
+        # # response.status = 409   # conflict
+        # # return {"error": f"A script is already running."}
+        # try:
+        #     pid = os.getpgid(running_script['process'].pid)
+        #     os.killpg(pid, signal.SIGTERM)  # Send the signal to all the process groups
+        #     # s = running_script['application']
+        #     running_script = None
+        #     # return {"message": f"application {category}/{application} stopped"}
+        # except Exception as e:
+        #     response.status = 500
+        #     return {"error": str(e)}
+        # stop_application(category, application)
+        script_path = os.path.join(app.config['scripts_dir'], running_script['category'], running_script['application']) + APP_SCRIPT_STOP_SUFFIX + APP_SCRIPT_EXT
+        if not os.path.isfile(script_path):
+            logging.error(f"unable to stop running application {script_path}, stop script not found")
+            response.status = 400
+            return {"error": f"unable to stop running application {running_script['category']}/{running_script['application']}, stop script not found"}
         try:
-            pid = os.getpgid(running_script['process'].pid)
-            os.killpg(pid, signal.SIGTERM)  # Send the signal to all the process groups
-            s = running_script['script']
-            running_script = None
-            # return {"message": f"script {s} stopped"}
+            # subprocess.Popen is non-blocking
+            # The os.setsid() is passed in the argument preexec_fn so
+            # it's run after the fork() and before exec() to run the shell.
+            logging.debug(f'calling {script_path}')
+            p = subprocess.Popen(script_path, stdout=subprocess.PIPE, shell=False, preexec_fn=os.setsid)
+            # return {"message": f"process {os.getpgid(p.pid)} called"}
         except Exception as e:
+            logging.exception(f'error when calling {script_path}')
             response.status = 500
             return {"error": str(e)}
 
-    script_path = os.path.join(app.config['scripts_dir'], category, script)
+    # logging.debug(f"scripts_dir {app.config['scripts_dir']}")
+    script_path = os.path.join(app.config['scripts_dir'], category, application) + APP_SCRIPT_START_SUFFIX + APP_SCRIPT_EXT
+    # logging.debug(f"script_path {script_path}")
     if not os.path.isfile(script_path):
+        logging.error(f"{script_path} start script not found")
         response.status = 400
-        return {"error": f"{category}/{script} not found"}
+        return {"error": f"{category}/{application} start script not found"}
     try:
         # subprocess.Popen is non-blocking
         # The os.setsid() is passed in the argument preexec_fn so
         # it's run after the fork() and before exec() to run the shell.
+        logging.debug(f'calling {script_path}')
         p = subprocess.Popen(script_path, stdout=subprocess.PIPE, shell=False, preexec_fn=os.setsid)
         running_script = {
             'category': category,
-            'script': script,
+            'application': application,
             'process': p
         }
-        return {"message": f"{category}/{script} started"}
+        return {"message": f"{category}/{application} started"}
     except Exception as e:
+        logging.exception(f'error when calling {script_path}')
         response.status = 500
         return {"error": str(e)}
 
 
-@app.route('/api/stop/<category>/<script>')
-def stop_script(category, script):
+@app.route('/api/stop/<category>/<application>')
+def stop_application(category, application):
     """
     https://stackoverflow.com/questions/4789837/how-to-terminate-a-python-subprocess-launched-with-shell-true/4791612#4791612
 
     :param category:
-    :param script: the filename of the script, without the path
+    :param application: the filename of the script, without the path
     :return:
     """
-    # TODO: rename "callScript"
-    script_path = os.path.join(app.config['scripts_dir'], category, script)
+    global running_script
+    script_path = os.path.join(app.config['scripts_dir'], category, application) + APP_SCRIPT_STOP_SUFFIX + APP_SCRIPT_EXT
     if not os.path.isfile(script_path):
+        logging.error(f"{category}/{application} start script not found")
         response.status = 400
-        return {"error": f"{category}/{script} does not exist"}
+        return {"error": f"{category}/{application} start script not found"}
     try:
         # subprocess.Popen is non-blocking
         # The os.setsid() is passed in the argument preexec_fn so
         # it's run after the fork() and before exec() to run the shell.
+        logging.debug(f'calling {script_path}')
         p = subprocess.Popen(script_path, stdout=subprocess.PIPE, shell=False, preexec_fn=os.setsid)
+        running_script = None
         return {"message": f"process {os.getpgid(p.pid)} called"}
     except Exception as e:
+        logging.exception(f'error when calling {script_path}')
         response.status = 500
         return {"error": str(e)}
 
@@ -159,12 +199,22 @@ def stop_current_script():
         pid = os.getpgid(running_script['process'].pid)
         os.killpg(pid, signal.SIGTERM)  # Send the signal to all the process groups
         c = running_script['category']
-        s = running_script['script']
+        s = running_script['application']
         running_script = None
         return {"message": f"script {c}/{s} stopped"}
     except Exception as e:
         response.status = 500
         return {"error": str(e)}
+
+
+@app.route('/api/stop')
+def panic_stop():
+    """
+    Call all the stop script one after the other.
+
+    TODO: create a panic_stop script ?
+    """
+    pass
 
 
 @app.route('/api/running')
@@ -176,7 +226,7 @@ def running():
     if running_script is not None:
         return {"running": {
             'category': running_script['category'],
-            'script': running_script['script']
+            'application': running_script['application']
         }}
     else:
         return {"running": None}
@@ -188,27 +238,45 @@ def ping():
     return {"message": "pong"}
 
 
-@app.route('/api/server/restart')
+# @app.route('/api/server/restart/<code>')
 def server_restart():
-    script_path = os.path.join(app.config['scripts_dir'], 'server', 'server_restart.sh')
+    # TODO: replace the code by a more robust security check; the code can easily be grabbed from the browser console.
+    logging.info(f'server_restart()')
+    # if code != '424242':
+    #     logging.debug(f"server_restart: invalid code")
+    #     return
+    script_path = os.path.join(app.config['scripts_dir'], 'system', 'server_restart.sh')
     if not os.path.isfile(script_path):
+        logging.error(f"{script_path} not found")
         response.status = 400
-        return {"error": f"{script_path} does not exist"}
+        return {"error": f"{script_path} not found"}
     try:
+        logging.debug(f'calling {script_path}')
         p = subprocess.Popen(script_path, stdout=subprocess.PIPE, shell=False, preexec_fn=os.setsid)
-        return {"message": f"process {os.getpgid(p.pid)} called"}
+        return {"message": f"restart started"}
     except Exception as e:
+        logging.exception(f'error when calling {script_path}')
         response.status = 500
         return {"error": str(e)}
 
 
-@app.route('/api/system/reboot')
-def system_reboot():
+@app.route('/api/system/reboot/<code>')
+def system_reboot(code):
+    # TODO: replace the code by a more robust security check; the code can easily be grabbed from the browser console.
+    logging.info(f'server_restart()')
+    if code != '424242':
+        logging.debug(f"system_reboot: invalid code")
+        return
     os.system('sudo reboot')
 
 
-@app.route('/api/system/poweroff')
-def server_poweroff():
+@app.route('/api/system/poweroff/<code>')
+def server_poweroff(code):
+    # TODO: replace the code by a more robust security check; the code can easily be grabbed from the browser console.
+    logging.info(f'server_poweroff()')
+    if code != '424242':
+        logging.debug(f"server_restart: invalid code")
+        return
     os.system('sudo poweroff')
 
 
@@ -240,13 +308,13 @@ def enable_cors():
 
 
 def setup_logging(log_destination):
-    if log_destination == '--':
+    if log_destination == '-':
         # Logging to standard output
-        logging.basicConfig(stream=sys.stdout, level=logging.INFO,
+        logging.basicConfig(stream=sys.stdout, level=logging.DEBUG if verbose else logging.INFO,
                             format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
     elif log_destination:
         # Logging to the specified file
-        logging.basicConfig(filename=log_destination, level=logging.INFO,
+        logging.basicConfig(filename=log_destination, level=logging.DEBUG if verbose else logging.INFO,
                             format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
     else:
         # Disable logging if no argument is passed
@@ -269,8 +337,8 @@ if __name__ == '__main__':
     parser.add_argument('-c', '--config', type=str, default=None, help='Path to the configuration file')
     parser.add_argument('-s', '--script', type=str, default=None, help='Name of the script to run')
     parser.add_argument('-v', '--verbose', action='store_true', help='Enable verbose mode')
-    parser.add_argument("--logfile", type=str, nargs='?', const='--', default=None,
-                        help="Specify '--' to log to standard output, or provide a filename to log to that file")
+    parser.add_argument("--logfile", type=str, nargs='?', const='-', default=None,
+                        help="Specify '-' to log to standard output, or provide a filename to log to that file")
     args = parser.parse_args()
 
     verbose = args.verbose
